Refactored the code according to new WLF best practice
