+++ webviz_subsurface/plugins/_well_completion/_utils/__init__.py
@@ -1,0 +1,1 @@
+from ._well_completion_data_model import WellCompletionDataModel
