+++ webviz_subsurface/plugins/_well_completion/_views/__init__.py
@@ -1,0 +1,5 @@
+from ._well_completion_view import (
+    ViewSettings,
+    WellCompletionView,
+    WellCompletionViewElement,
+)
