+++ webviz_subsurface/plugins/_well_completion/_views/_well_completion_view/__init__.py
@@ -1,0 +1,2 @@
+from ._view import ViewSettings, WellCompletionView
+from ._view_element import WellCompletionViewElement
