+++ webviz_subsurface/plugins/_well_completion/_views/_well_completion_view/_view_element.py
