--- webviz_subsurface/plugins/_well_completion/view_elements/__init__.py
@@ -1,1 +1,0 @@
-from ._well_completion_view_element import WellCompletionViewElement
