--- webviz_subsurface/plugins/_well_completion/views/__init__.py
@@ -1,1 +1,0 @@
-from ._well_completion_view import WellCompletionView
